backend/requirements.txt
fastapi
uvicorn
pillow
torch
transformers
python-multipart
pytesseract
